Smz's Update for 9/1/15
Adds the ignore elemental text for those on which it would fit for.
Changed some modifiers on most enemies, removed elements from npcs
weapons, and removes str. spd, and def down from the ignitions that had
them.

Also changing Torcleaver to not be magic and making its AoE damage and
TP cost to match the -jaII spells as well