Slight Titanite price changes.
Doing a slight price lower on the +1 and +5, with a slight increase to
the one Beedle carries. Even though the Beedle one would technically be
3 million, to match 40 of the +5 ones, decided to not be that accurate
with it. Also did a slight change on the spoils for Freya and Sif, now
that I realize how spoils actually work. This should potentially make it
easier to get both seals.