Added WonderGuard flag for monsters/bosses
In [info] put WonderGuard=true to be immune to anything that doesn't
have a value or a value set to 0 in modifiers.

The easiest way to make a monster immune to everything except 1 thing
would be to only put the 1 thing in modifiers and erase the rest out of
that section.