New 1 orb cost weapons, new monsters.
Added a variety of basic weapons that cost one black orb and have one
tech that can be bought.

New monsters that are only weak to certain weapons and their techs.
Streak 250+ for most, streak 350+ for another for now on these.

Some changes as well for a potential high level "super boss."