April 1st update (Legit, not fake.)
A lot of changes. Not quite sure exactly, but I'll try to list them all.

Updated modifiers on the new monsters, an update to Lirarian as well.
New torment weapons, new gold voucher weapons that use ammo. the new
shotguns are double barrels that fire both at once, so they need 2 ammo
per shot.  due to that, the ammo is half the cost compared to the arrow
or bullet ammo. The techs will also use up the ammo, so make sure you
don't run out.

Fixed some typos that I saw on certain techs and ignition descs.

Rokcet Launcher is on the auction house, at a "high" cost. Changed
Hero's Medal to be a player version, which requires a new torment sword
to get use out of. New accessory for Frog and the Cena npc, new taunts
for Bayonetta along with some adjustments.

Changed the event on the Iron Keep battlefield to target players instead
of random, which would occasionally target the boss.