August 1, 2016 Update
* Added the Supply Wagon battle type.  In this type players will need to
protect a wagon against monsters.  If the wagon dies or all players die
the battle fails.  Players will be level synced to whatever the current
battlestreak is for this battle.  Trusts and Summons are allowed though,
to help.

* Fixed a serious bug that was preventing accessories with more than one
type from working (thanks to Andrio and Pentium320 for finding and
helping me fix it)