Most Active Skills Must Be Equipped To Use!
Starting with this update most of the active skills must be equipped in
order to be used.  There are exceptions (enhancing point and style
skills do not, as well as a few others).

To equip a skill: !skills equip skillname
To unequip a skill: !skills unequip skillname
To check: !skills equipped

Active skills that are equipped will also have a [] around the name in
the skill list.

Everyone has 5 skill slots available to them to equip what they need for
battle.