The Witch update, ignoreTP flag on others
Death condition for The Witch in Torment should now be two things, which
it'll remain the same outside of Torment. Actually planned to have this
in another update, but I'll do it now.

Added the IgnoreTP flag for other bosses due to using player techs.

Fixed a typo on the desc for another boss.