Added enmity system
Added an enmity targeting system.  Basically attacking monsters and
healing damage will generate enmity and the monsters will target the
player who has the highest enmity.  If monsters go first or if it's a
supply wagon/allied president battle the monsters will randomly pick a
target like before.

Also added the !toggle enmity  to turn this on/off.  Note that it's on
by default but only if the bot restarts and the option is added to the
system.dat.  Otherwise it will not be on (a bot admin will need to use
the command).