June 29, 2017 update
SlimeOil moved from "misc" item to "battle" item.  It now inflicts the
"heavy" status to a target.

Heavy status added.  When applied to a target the target will no longer
be able to fly for the duration of the battle.  If the target is flying
when the status is applied, the target will be sent back to the ground.