Dragon Hunt Battle Level is now different from Age
It's now possible for dragon hunt battles to have independent battle
levels instead of always being the dragon's age.  This means it is now
possible for dragons to become stronger than the players the more times
players die to the dragon as the players will always be synced to the
dragon's initial age even if the dragon has leveled up since then.

To counter this, a special shop purchase will be made available soon
that will either increase the battle level or to decrease dragon's ages.