In cosmic battles only right hand weapon techs can be used
With this update players can only use techniques that are performed with
the right-handed weapon if they are in the cosmic battle type.  This is
to fix an error where players could previously equip a legendary in the
right hand but use techniques from a non-legendary weapon in the left
hand.

Fixes #48