French Taunter and misc updates
Changing the French Taunter npc to actually attack so that AI vs AI battles don't potentially drag on for 30 minutes or more.

Changes to the Spectator enemy, so maybe it's not so strong on outpost battles.

Potentially increasing the regen potency for Bayonetta if she uses the Umbran ignition.

Fixing the color on the Asriel boss quote to normal after the name.