March 14, 2018 update
* Added a new style: HIGH ROLLER.  This style is a gambling/luck based.
While using this style you can use the !gamble skill for 200 orbs
(instead of 1000) and without equipping that skill.  You also have
access to the Card Shark skill that can unleash the powers of monsters
captured in trading cards.  To use it, use !cardshark tradingcard-#
target    The technique used is random and is based on the monster
pictured on the card.

* Added a new monster type: Hippogryph.  Also included a names list and
a new boss of that type.