June 18, 2018 Update
* Added the !enmity command.  This command will show who the monsters
are targeting.

* Added new monsters

* Added the LucidDreaming active skill. This skill will allow you to cut
your enmity in half.  It cannot be bought and must be learned.

* Added the SoftBlows active skill.  This skill allows your next attack
to generate no enmity.  It cannot be bought and must be learned.

* Added the corresponding skill learning items (not currently no way to
obtain them)

KNOWN ISSUE: When the baby hydra is summoned into the battle it will
show the wrong # of heads in its description.  It is set to 2, not 1.