August 30, 2018 Update
* Added the DESYNTH skill. Players who have purchased the Alchemy skill
at least once will be able to purchase the desynth skill. This skill
allows players to break a crafted item back into their individual
ingredients.  Players will receive 1 of the ingredients back.

* Changed food type items so that it's now possible for them to affect
more than one stat at a time.  While hp and tp can be included it is
recommended that they are not due to the caps on both of those and an
error message that will occur if the cap is hit.