September 10, 2018 Update
* Added the TREASUREHUNTER style. This style will enhance dagger usage,
increase the player's chances of receiving spoils, enhance the steal
skill and give players access to the !trickattack and !sneakattack
skills.

* Added the TrickAttack skill. This skill can only be used while using
the TreasureHunter style.  If a player has the most enmity in battle
this skill will guarantee the next hit will be a direct critical attack.
After a successful hit, the player's enmity will be reset to 1.

* Added the SneakAttack skill. This skill can only be used while using
the TreasureHunter style.  If a player has the least enmity in battle
this skill will guarantee the next hit will be a direct critical attack.

* Added the doom status effect into the game

* Changed the Trickster style so that it no longer affects the steal
skill as this has been moved to the TreasureHunter style.

* Changed !view-help skill command so that the style-based skills (such
as TrickAttack, PerfectCounter, Wrestle, etc) can now be seen if
searched.

* Changed the !enmity command so that it will show who has the lowest in
addition to the highest enmity in battle if there is more than one
player in battle and at least two people are on the enmity list.