December 4, 2019 Update
Added the ability for str, int and def to "level up" on their own as characters play the game.  STR will increase with melee attacks, INT with techs and DEF with getting attacked by monsters.  10 stat xp = +1 to the stat.

Fixed an issue with !view-info tech that was showing the red orb cost in kill coin mode.

Changed the amount of orbs newer players get slightly and made the lower orb tiers a little less harsh.