March 30, 2020 Update
ADDED Beastmen Tribe Enmity. When certain beastmen are killed in battle it will increase their enmity. When the enmity hits a certain threshold there is a chance that a primal will be summoned and a special dungeon unlocked.

Check beastmen enmity with !beastment enmity [type]   where [type] is Orc, Lamia, Yagudo or Quadav

ADDED a new Treasure Hunter style skill: Bribe.  Using this players can attempt to bribe a monster with red orbs to leave the battle.  Use !bribe monstername amountoforbs     The higher number of orbs offered the better odds that the monster will accept.  This command only works on normal battles and the higher streaks will cost more orbs to bribe the same monsters.

FIXED a few minor errors here and there.