This update must be the work of an enemy stand. (September 9th, 2020)
This update was originally going to be a few bosses, but ended up adding more and fixing some other issues.

Updated all of the "new_boss" and such files to have more information on certain things compared to before.

So basically, just several new bosses, and a new monster. Fixed some issues on two monsters. Added Elite tags to all of the npcs, except for three. Made the overpowered Rare Akuma npc even more overpowered as it's even less of a chance that you'll see this npc now.