The start of discord command recognition
This is a serious work in progress and may not be finished any time soon.

This is a way for the bot to recognize commands coming from a discord-irc bridge bot.  Note that the bridge's text has to be formatted a specific way for BattleArena to recognize commands from it.   Also in system.dat disable colors with AllowColors=false to make the outgoing text read better on discord.