August 29, 2021 Update
Shido Pollendina the engineer NPC added.  He will show up when the garden is level 10 and there's at least 1 player who is level 1000.   He sells RocketFuel and OxygenTank items in the misc shop.

These two items will later be used to craft a dungeon key.  However the dungeons are not finished so as of now they do not do anything.