May 18, 2020 Update
Pirate battles are now a dungeon that can randomly be started when a new battle begins as long as the streak is 50+.   Right now it is a 1.9% chance of being selected.

The dungeon is 3 battles--two against minions and one against a boss. There is also a new boss to go along with the old SilverHook and BlackBeard bosses.