December 6 Adjustments
Changed the recipe for the DensetsuNoYoroi+2 to require 50 RefinedLunarOre instead of 100.

Changed the dungeon rest timer so that it is now based on active players. If there are 4+ active players then a player can start a dungeon every 4 hours.  If there are less than 4 then a player can start a dungeon every 2 hours.