December 9, 2022 Update
ADDED the Echo to certain dungeons. If players lose on a battle the echo will reset the battle and give them another try.

CHANGED the Frost Monster battle type to be a dungeon that is only available in December.  Buy a FrostOrb from the dungeon key shop for 1 BeastmenSeal to gain access to it.  The first time the dungeon is cleared Santa will be saved. Each time afterwards an elf will be saved.

FIXED yet another bug with mini.